feat: completed functionality for getting form data into the array
To get the data from the form, I used a submit event listener to create
a FormData object that contains the values of the form. I then created a
formdata event listener that takes that new form object and iterates it
into a new array that then gets pushed into the addBookToLibrary
function.